+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - WEEKLY PLANNING : Add a type and minor changes
 - Production : Remove stockLocation from Machine Type object.
 - SupplyChain : Addition of multiselect field in MrpLineType object.
+- MANUFACTURING : New object S&OP (PIC)
 
 ## Improvements
 - SaleOrder: refactored 'Business Project' panel